fix(frontend): check issue button
