+++ frontend/app/src/components/history/events/HistoryEventsView.vue
@@ -80,7 +80,6 @@ const toggles = ref<HistoryEventsToggles>({
 });
 
 const showAlerts = ref<boolean>(false);
-const manualIssueCheck = ref<boolean>(false);
 const currentAction = ref<HistoryEventAction>(HISTORY_EVENT_ACTIONS.QUERY);
 
 const dialogContainer = useTemplateRef<InstanceType<typeof HistoryEventsDialogContainer>>('dialogContainer');
@@ -236,12 +235,10 @@ watchDebounced(route, async () => {
       <template #buttons>
         <HistoryEventsViewButtons
           v-model:show-alerts="showAlerts"
-          v-model:manual-issue-check="manualIssueCheck"
           :processing="processing"
           :loading="anyEventsDecoding"
           :include-evm-events="includes.evmEvents"
           :main-page="mainPage"
-          :section-loading="debouncedProcessing"
           @refresh="actions.refresh.all(true, $event)"
           @show:dialog="dialogContainer?.show($event)"
         />
@@ -252,7 +249,6 @@ watchDebounced(route, async () => {
           v-model:show="showAlerts"
           :processing="processing"
           :main-page="mainPage"
-          :manual-issue-check="manualIssueCheck"
           @open:match-asset-movements="openMatchAssetMovementsDialog()"
         />
